+++ lib/openapi_first/response_header.rb
@@ -5,5 +5,5 @@ require_relative 'parameter'
 module OpenapiFirst
   # A header of a response definition.
   # @attr_reader [Parameter] parameter The header as a Parameter, which knows how to unpack a raw value.
-  Header = Data.define(:name, :required?, :schema, :parameter)
+  ResponseHeader = Data.define(:name, :required?, :schema, :parameter)
 end