+++ lib/openapi_first/validators/request_body.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require_relative 'multipart_request_body'
-require_relative 'required_request_body'
 
 module OpenapiFirst
   module Validators
@@ -9,9 +8,14 @@ module OpenapiFirst
       MULTIPART = %r{\Amultipart/}i
       private_constant :MULTIPART
 
+      REQUIREMENT = lambda do |parsed_request|
+        Failure.new(:invalid_body, message: 'Request body must not be empty') if parsed_request.body.nil?
+      end
+      private_constant :REQUIREMENT
+
       def self.for(content_schema:, required_request_body:, content_type:)
         validators = []
-        validators << RequiredRequestBody.new if required_request_body
+        validators << REQUIREMENT if required_request_body
         klass = MULTIPART.match?(content_type.to_s) ? MultipartRequestBody : self
         validators << klass.new(content_schema:)
         validators